Handle conditional fsevents import
