+++ src/watch/fsevents-importer.ts
@@ -1,0 +1,18 @@
+let fsEvents: any;
+let fsEventsImportError: any;
+
+export function loadFsEvents() {
+	return import('fsevents')
+		.then(namespace => {
+			fsEvents = namespace;
+		})
+		.catch(err => {
+			fsEventsImportError = err;
+		});
+}
+
+// A call to this function will be injected into the chokidar code
+export function getFsEvents() {
+	if (fsEventsImportError) throw fsEventsImportError;
+	return fsEvents;
+}
